util/timeutil: Move util/timer.go to timeutil package
