+++ kv/txn_coord_sender.go
@@ -36,6 +36,7 @@ import (
 	"github.com/cockroachdb/cockroach/util/log"
 	"github.com/cockroachdb/cockroach/util/metric"
 	"github.com/cockroachdb/cockroach/util/stop"
+	"github.com/cockroachdb/cockroach/util/timeutil"
 	"github.com/cockroachdb/cockroach/util/tracing"
 	"github.com/cockroachdb/cockroach/util/uuid"
 )
@@ -194,7 +195,7 @@ func NewTxnCoordSender(
 // stats).
 func (tc *TxnCoordSender) startStats() {
 	res := time.Millisecond // for duration logging resolution
-	var statusLogTimer util.Timer
+	var statusLogTimer timeutil.Timer
 	defer statusLogTimer.Stop()
 	scale := metric.Scale1M
 	for {
